Revert "feat: refactor lib->src (#16)"
This reverts commit b84e5a92fd519f09944052e510849ee37de96dd4.
