+++ lib/chai.js
