+++ lib/matcherGenerator.js
