+++ lib/toMatchEyesScreenshot.js
