+++ package.json
@@ -17,7 +17,7 @@
   },
   "jest": {
     "preset": "jest-puppeteer",
-    "setupTestFrameworkScriptFile": "<rootDir>/src/jest.js"
+    "setupTestFrameworkScriptFile": "<rootDir>/lib/jest.js"
   },
   "devDependencies": {
     "chai": "^4.1.2",
